+++ modules/core/system.nix
@@ -10,8 +10,27 @@ in {
         "nix-command"
         "flakes"
       ];
-      substituters = ["https://hyprland.cachix.org"];
-      trusted-public-keys = ["hyprland.cachix.org-1:a7pgxzMz7+chwVL3/pzj6jIBMioiJM7ypFP8PwtkuGc="];
+      substituters = [
+        "https://cache.nixos.org?priority=10"
+        "https://nix-config.cachix.org"
+        "https://nix-community.cachix.org"
+        "https://hyprland.cachix.org"
+        "https://flyer-myer.cachix.org"
+      ];
+      trusted-public-keys = [
+        "cache.nixos.org-1:6NCHdD59X431o0gWypbMrAURkbJ16ZPMQFGspcDShjY="
+        "nix-config.cachix.org-1:Vd6raEuldeIZpttVQfrUbLvXJHzzzkS0pezXCVVjDG4="
+        "nix-community.cachix.org-1:mB9FSh9qf2dCimDSUo8Zy7bkq5CX+/rkCWyvRCYg3Fs="
+        "hyprland.cachix.org-1:a7pgxzMz7+chwVL3/pzj6jIBMioiJM7ypFP8PwtkuGc="
+        "flyer-myer.cachix.org-1:giZCeviY2BgmVUxFYvPwpF7Pim68/gWpL9BbqPGiS54="
+      ];
+
+      # Add trusted-users setting
+      trusted-users = ["root" "helloworld74"];
+
+      # for direnv GC roots
+      keep-derivations = true;
+      keep-outputs = true;
     };
   };
   time.timeZone = "Asia/Singapore";
@@ -29,7 +48,7 @@ in {
   };
   environment.variables = {
     NIXOS_OZONE_WL = "1";
-    GGOS_VERSION = "2.6.1";
+    GGOS_VERSION = "2.6.3";
     GGOS = "true";
   };
   console.keyMap = "${consoleKeyMap}";