+++ modules/home/terminals/rio.nix
@@ -3,5 +3,20 @@
 in {
   programs.rio = {
     enable = true;
+    settings = {
+      "confirm-before-quit" = false;
+      fonts.family = lib.mkForce fontFamily;
+      window = {
+        opacity = lib.mkForce 0.90;
+        opacity-cells = false;
+      };
+      bindings.keys = [
+        {
+          key = "q";
+          "with" = "super";
+          action = "Quit";
+        }
+      ];
+    };
   };
 }